requirements.txt
aiohappyeyeballs              
aiohttp                       
aiohttp-retry                 
aiosignal                     
asgiref                       
attrs                         
certifi                       
charset-normalizer            
Django                        
django-cors-headers           
djangorestframework           
djangorestframework_simplejwt 
frozenlist                    
idna                          
multidict                     
mysqlclient                   
pillow                        
pip                           
propcache                     
PyJWT                         
requests                      
sqlparse                      
twilio                       
tzdata                       
urllib3                      
yarl                          
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary